+++ mmpm/api/endpoints/index.py
@@ -4,9 +4,9 @@ from flask import Blueprint, Response, jsonify
 
 from mmpm.api.constants import http
 from mmpm.api.endpoints.endpoint import Endpoint
-from mmpm.log.logger import MMPMLogger
+from mmpm.log.factory import MMPMLogFactory
 
-logger = MMPMLogger.get_logger(__name__)
+logger = MMPMLogFactory.get_logger(__name__)
 
 
 class Index(Endpoint):