+++ mmpm/api/repeater.py
@@ -16,9 +16,9 @@ from time import sleep
 import socketio
 
 from mmpm.env import MMPMEnv
-from mmpm.log.logger import MMPMLogger
+from mmpm.log.factory import MMPMLogFactory
 
-logger = MMPMLogger.get_logger(__name__)
+logger = MMPMLogFactory.get_logger(__name__)
 
 
 def create():