+++ mmpm/subcommands/_sub_cmd_completion.py
@@ -2,10 +2,10 @@
 """ Command line options for 'completion' subcommand """
 from typing import Dict
 
-from mmpm.log.logger import MMPMLogger
+from mmpm.log.factory import MMPMLogFactory
 from mmpm.subcommands.sub_cmd import SubCmd
 
-logger = MMPMLogger.get_logger(__name__)
+logger = MMPMLogFactory.get_logger(__name__)
 
 
 class Completion(SubCmd):