+++ mmpm/subcommands/_sub_cmd_mm_ctl.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python3
 """ Command line options for 'mm-ctl' subcommand """
 from mmpm.env import MMPMEnv
-from mmpm.log.logger import MMPMLogger
+from mmpm.log.factory import MMPMLogFactory
 from mmpm.magicmirror.controller import MagicMirrorController
 from mmpm.magicmirror.magicmirror import MagicMirror
 from mmpm.subcommands.sub_cmd import SubCmd
 from mmpm.utils import confirm
 
-logger = MMPMLogger.get_logger(__name__)
+logger = MMPMLogFactory.get_logger(__name__)
 
 
 class MmCtl(SubCmd):