+++ mmpm/ui.py
@@ -1,5 +1,11 @@
 #!/usr/bin/env python3
-import importlib.resources as pkg_resources
+import sys
+
+if sys.version_info < (3, 9):
+    import importlib_resources as pkg_resources
+else:
+    import importlib.resources as pkg_resources
+
 import json
 import os
 import shutil
@@ -7,11 +13,11 @@ from pathlib import Path
 
 from mmpm.__version__ import version
 from mmpm.constants import urls
-from mmpm.log.logger import MMPMLogger
+from mmpm.log.factory import MMPMLogFactory
 from mmpm.singleton import Singleton
 from mmpm.utils import run_cmd
 
-logger = MMPMLogger.get_logger(__name__)
+logger = MMPMLogFactory.get_logger(__name__)
 
 
 class MMPMui(Singleton):